test(core): move lookup-subject dedup specs into shared base spec
The recursive-lookup id-list dedup happens in the shared LookupSubjectEngine
before the provider call, so the regression and its performance invariant are
provider-agnostic. Move both specs out of the InMemory-only suite into
BaseLookupSubjectEngineSpecs so they also run against the SQL Server and
Postgres providers.

The capturing decorator is now wired generically: DecorateDataReaderToCapture
wraps whichever IDataReaderProvider the active provider registered, instead of
resolving the concrete InMemoryProvider, so the same id-list invariant is
exercised across all three backends.
