+++ tests/Valtuutus.Tests.Shared/BaseLookupSubjectEngineSpecs.cs
@@ -4,10 +4,12 @@ using Valtuutus.Core.Schemas;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using Valtuutus.Core.Configuration;
+using Valtuutus.Core.Engines.LookupEntity;
 using Valtuutus.Core.Engines.LookupSubject;
 using Valtuutus.Data;
 using Valtuutus.Core.Data;
 using Valtuutus.Core.Lang;
+using Valtuutus.Core.Pools;
 
 namespace Valtuutus.Tests.Shared;
 
@@ -22,12 +24,17 @@ public abstract class BaseLookupSubjectEngineSpecs : IAsyncLifetime
 
     protected IDatabaseFixture Fixture { get; }
 
-    private ServiceProvider CreateServiceProvider(string? schema = null)
+    private ServiceProvider CreateServiceProvider(string? schema = null, List<List<string>>? captures = null)
     {
         var services = new ServiceCollection()
             .AddValtuutusCore(schema ?? TestsConsts.DefaultSchema);
         AddSpecificProvider(services)
             .AddConcurrentQueryLimit(3);
+
+        if (captures is not null)
+        {
+            DecorateDataReaderToCapture(services, captures);
+        }
 
         return services.BuildServiceProvider();
     }
@@ -43,6 +50,46 @@ public abstract class BaseLookupSubjectEngineSpecs : IAsyncLifetime
         var provider = scope.ServiceProvider.GetRequiredService<IDataWriterProvider>();
         await provider.Write(tuples, attributes, default);
         return lookupSubjectEngine;
+    }
+
+    /// <summary>
+    /// Builds an engine like <see cref="CreateEngine"/> but wraps the configured provider's
+    /// <see cref="IDataReaderProvider"/> so every entity-id list the engine forwards into a
+    /// recursive query is recorded. Resolution behaviour stays real — only the forwarded ids are
+    /// observed — and the decoration is provider-agnostic, so the same invariant is exercised
+    /// against the in-memory, SQL Server, and Postgres providers.
+    /// </summary>
+    private async ValueTask<(ILookupSubjectEngine Engine, List<List<string>> Captures)> CreateEngineWithCapture(
+        RelationTuple[] tuples, AttributeTuple[] attributes, string? schema = null)
+    {
+        var captures = new List<List<string>>();
+        var serviceProvider = CreateServiceProvider(schema, captures);
+        var scope = serviceProvider.CreateScope();
+        var lookupSubjectEngine = scope.ServiceProvider.GetRequiredService<ILookupSubjectEngine>();
+        var provider = scope.ServiceProvider.GetRequiredService<IDataWriterProvider>();
+        await provider.Write(tuples, attributes, default);
+        return (lookupSubjectEngine, captures);
+    }
+
+    private static void DecorateDataReaderToCapture(IServiceCollection services, List<List<string>> captures)
+    {
+        // Wrap whatever provider registered IDataReaderProvider rather than mock the interface, so
+        // resolution behaviour stays real and only the forwarded entity-id lists are observed.
+        var original = services.Last(d => d.ServiceType == typeof(IDataReaderProvider));
+        services.Remove(original);
+        services.Add(ServiceDescriptor.Describe(
+            typeof(IDataReaderProvider),
+            sp => new CapturingReader(CreateInner(sp, original), captures),
+            original.Lifetime));
+    }
+
+    private static IDataReaderProvider CreateInner(IServiceProvider sp, ServiceDescriptor original)
+    {
+        if (original.ImplementationInstance is IDataReaderProvider instance)
+            return instance;
+        if (original.ImplementationFactory is not null)
+            return (IDataReaderProvider)original.ImplementationFactory(sp)!;
+        return (IDataReaderProvider)ActivatorUtilities.CreateInstance(sp, original.ImplementationType!);
     }
 
 
@@ -351,5 +398,148 @@ public abstract class BaseLookupSubjectEngineSpecs : IAsyncLifetime
         // alice is owner → excluded; bob is editor (exists in relation_tuples) but not owner → included
         result.Should().Contain("bob");
         result.Should().NotContain("alice");
+    }
+
+    // Two-hop tuple-to-userset chain (workspace.view := groups.members → group.members := teams.member)
+    // so a duplicate intermediate entity appears at an *intermediate* forwarding step — exactly where
+    // the id-list bloat lived.
+    private const string DedupSchema = """
+        entity user {}
+        entity team {
+            relation member @user;
+        }
+        entity group {
+            relation teams @team;
+            permission members := teams.member;
+        }
+        entity workspace {
+            relation groups @group;
+            permission view := groups.members;
+        }
+        """;
+
+    // g1 and g2 both belong to t1, so resolving `teams` for [g1, g2] yields the team id twice.
+    private static readonly RelationTuple[] ConvergingTuples =
+    [
+        new("workspace", "w1", "groups", "group", "g1"),
+        new("workspace", "w1", "groups", "group", "g2"),
+        new("group", "g1", "teams", "team", "t1"),
+        new("group", "g2", "teams", "team", "t1"),
+        new("team", "t1", "member", "user", "alice"),
+    ];
+
+    [Fact]
+    public async Task Lookup_WithConvergingIntermediateEntities_ReturnsDistinctSubjects()
+    {
+        // Arrange
+        var engine = await CreateEngine(ConvergingTuples, [], DedupSchema);
+
+        // Act
+        var result = await engine.Lookup(
+            new LookupSubjectRequest("workspace", "view", "user", "w1"),
+            CancellationToken.None);
+
+        // Assert — alice is reachable through both g1 and g2, but must appear exactly once.
+        result.Should().BeEquivalentTo("alice");
+    }
+
+    [Fact]
+    public async Task Lookup_DoesNotForwardDuplicateEntityIds_ToRecursiveLookups()
+    {
+        // Arrange — capture every entity-id list the engine forwards into a recursive query.
+        var (engine, captures) = await CreateEngineWithCapture(ConvergingTuples, [], DedupSchema);
+
+        // Act
+        var result = await engine.Lookup(
+            new LookupSubjectRequest("workspace", "view", "user", "w1"),
+            CancellationToken.None);
+
+        // Assert
+        result.Should().BeEquivalentTo("alice");
+
+        // Performance invariant: no recursion level is ever handed a list containing duplicates,
+        // even though `teams` for [g1, g2] resolves to t1 twice.
+        captures.Should().NotBeEmpty("the engine must have issued at least one entity-id query");
+        captures.Should().OnlyContain(
+            ids => ids.Distinct().Count() == ids.Count,
+            "duplicate ids bloat the TVP / id array without changing results");
+
+        // Specifically, the team -> member leaf query is fed the deduplicated [t1], not [t1, t1].
+        captures.Should().Contain(ids => ids.Count == 1 && ids[0] == "t1");
+    }
+
+    /// <summary>
+    /// Thin decorator that delegates every call to the real provider and records the entity-id
+    /// lists passed into <see cref="GetRelationsWithEntityIds"/>.
+    /// </summary>
+    private sealed class CapturingReader(IDataReaderProvider inner, List<List<string>> captures) : IDataReaderProvider
+    {
+        public Task<PooledList<RelationTuple>> GetRelationsWithEntityIds(EntityRelationFilter entityRelationFilter,
+            string subjectType, IEnumerable<string> entityIds, string? subjectRelation, CancellationToken cancellationToken)
+        {
+            var materialized = entityIds.ToList();
+            captures.Add(materialized);
+            return inner.GetRelationsWithEntityIds(entityRelationFilter, subjectType, materialized, subjectRelation,
+                cancellationToken);
+        }
+
+        public Task<SnapToken?> GetLatestSnapToken(CancellationToken cancellationToken)
+            => inner.GetLatestSnapToken(cancellationToken);
+
+        public Task<PooledList<RelationTuple>> GetRelations(RelationTupleFilter tupleFilter, CancellationToken cancellationToken)
+            => inner.GetRelations(tupleFilter, cancellationToken);
+
+        public Task<bool> HasDirectRelation(RelationTupleFilter tupleFilter, string subjectId, CancellationToken cancellationToken)
+            => inner.HasDirectRelation(tupleFilter, subjectId, cancellationToken);
+
+        public Task<bool> HasAnyDirectRelation(string entityType, string[] entityIds, string relation, string subjectId,
+            SnapToken snapToken, CancellationToken cancellationToken)
+            => inner.HasAnyDirectRelation(entityType, entityIds, relation, subjectId, snapToken, cancellationToken);
+
+        public Task<bool> HasTupleToUserSetRelation(string entityType, string entityId, string tupleSetRelation,
+            string subEntityType, string computedRelation, string subjectType, string subjectId, SnapToken snapToken,
+            CancellationToken cancellationToken)
+            => inner.HasTupleToUserSetRelation(entityType, entityId, tupleSetRelation, subEntityType, computedRelation,
+                subjectType, subjectId, snapToken, cancellationToken);
+
+        public Task<PooledList<RelationTuple>> GetIndirectRelations(RelationTupleFilter tupleFilter, CancellationToken cancellationToken)
+            => inner.GetIndirectRelations(tupleFilter, cancellationToken);
+
+        public Task<PooledList<RelationTuple>> GetRelationsWithSubjectsIds(EntityRelationFilter entityFilter,
+            string[] subjectsIds, string subjectType, EntityScope? scope, CancellationToken cancellationToken)
+            => inner.GetRelationsWithSubjectsIds(entityFilter, subjectsIds, subjectType, scope, cancellationToken);
+
+        public Task<PooledList<RelationTuple>> GetRelationsJoined(EntityRelationFilter mainFilter, string subEntityType,
+            string subRelation, string subjectType, string subjectId, EntityScope? scope, CancellationToken cancellationToken)
+            => inner.GetRelationsJoined(mainFilter, subEntityType, subRelation, subjectType, subjectId, scope, cancellationToken);
+
+        public Task<AttributeTuple?> GetAttribute(EntityAttributeFilter filter, CancellationToken cancellationToken)
+            => inner.GetAttribute(filter, cancellationToken);
+
+        public Task<List<AttributeTuple>> GetAttributes(EntityAttributeFilter filter, CancellationToken cancellationToken)
+            => inner.GetAttributes(filter, cancellationToken);
+
+        public Task<Dictionary<(string AttributeName, string EntityId), AttributeTuple>> GetAttributes(
+            EntityAttributesFilter filter, EntityScope? scope, CancellationToken cancellationToken)
+            => inner.GetAttributes(filter, scope, cancellationToken);
+
+        public Task<List<AttributeTuple>> GetAttributesWithEntityIds(AttributeFilter filter, IEnumerable<string> entitiesIds,
+            CancellationToken cancellationToken)
+            => inner.GetAttributesWithEntityIds(filter, entitiesIds, cancellationToken);
+
+        public Task<Dictionary<(string AttributeName, string EntityId), AttributeTuple>> GetAttributesWithEntityIds(
+            EntityAttributesFilter filter, IEnumerable<string> entitiesIds, CancellationToken cancellationToken)
+            => inner.GetAttributesWithEntityIds(filter, entitiesIds, cancellationToken);
+
+        public Task<PooledList<AttributeTuple>> GetAttributesSingleEntity(EntityAttributesFilter filter, CancellationToken cancellationToken)
+            => inner.GetAttributesSingleEntity(filter, cancellationToken);
+
+        public Task<List<string>> GetEntityIdsExcluding(string entityType, IReadOnlyCollection<string> excludeIds,
+            SnapToken snapToken, CancellationToken cancellationToken)
+            => inner.GetEntityIdsExcluding(entityType, excludeIds, snapToken, cancellationToken);
+
+        public Task<List<string>> GetSubjectIdsExcluding(string subjectType, IReadOnlyCollection<string> excludeIds,
+            SnapToken snapToken, CancellationToken cancellationToken)
+            => inner.GetSubjectIdsExcluding(subjectType, excludeIds, snapToken, cancellationToken);
     }
 }
